+++ src/slogo/model/backEndInternal/commands/IsPenDown.java
@@ -2,7 +2,7 @@ package slogo.model.backEndInternal.commands;
 
 import slogo.model.backEndInternal.BackEndTurtle;
 
-public class IsPenDown implements Command<Integer> {
+public class IsPenDown implements Command<Double> {
 
   BackEndTurtle backEndTurtle;
 
@@ -11,7 +11,7 @@ public class IsPenDown implements Command<Integer> {
   }
 
   @Override
-  public double execute() {
-    return backEndTurtle.getPenStatus() ? 0 : 1;
+  public Double execute() {
+    return (double) (backEndTurtle.getPenStatus() ? 0 : 1);
   }
 }
